requirements.txt
asgiref==3.8.1
Django==5.2
sqlparse==0.5.3
psycopg2-binary==2.9.3
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6
django-allauth==0.50.0
environs[django]==9.5.0
marshmallow==3.20.1
pillow==10.0.0
django-debug-toolbar==4.2.0
whitenoise>=6.1,<7
gunicorn>=20.1.0,<22.0
cloudinary
django-cloudinary-storage
django-redis

wordcloud
nltk
networkx
pandas
matplotlib
regex
xlrd
numpy
datetime

fuzzywuzzy
rapidfuzz
seaborn 
openai
tiktoken
openpyxl>=3.1.0


# AWS
boto3
django-storages
python-Levenshtein==0.23.0

# Heroku buildpack requirement
dj-database-url
